+++ packages/backend/convex/functions/auth.config.ts
@@ -1,6 +1,8 @@
-import { getAuthConfigProvider } from "@convex-dev/better-auth/auth-config"
 import type { AuthConfig } from "convex/server"
+import { getAuthConfigProvider } from "kitcn/auth/config"
 
 export default {
-	providers: [getAuthConfigProvider()],
+	providers: [
+		process.env.JWKS ? getAuthConfigProvider({ jwks: process.env.JWKS }) : getAuthConfigProvider(),
+	],
 } satisfies AuthConfig